Release should be called Source
If we work with UFO, they are meant for collaborators to built on.
They do not work for end users. Therefore, they are the source.

Uptil now most of the work has been done in FontForge.
For now it makes sense to keep the sfdir files in ‘Work’,
and generate UFO files from them.

Once collaborators start to use the UFO’s, we will have to find a way
to merge the changes back into the FontForge file.